:art: move container simple parsers into Container.js
